requirement_file.txt
asgiref==3.5.0
Django==4.0.2
sqlparse==0.4.2
djangorestframework==3.13.1